# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sistema de Gestão Financeira Pessoal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List categories
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Transactions
...
  /budgets:
    get:
      operationId: getBudgets
      tags: [budgets]
      summary: List budgets for a month
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create budget
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /budgets/status:
    get:
      operationId: getBudgetStatus
      tags: [budgets]
      summary: Budget vs actual spending by category
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Budget status list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetStatus"
  /budgets/{id}:
    put:
      operationId: updateBudget
      tags: [budgets]
      summary: Update budget
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete budget
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Goals
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Total balance, monthly income/expenses
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/by-category:
    get:
      operationId: getDashboardByCategory
      tags: [dashboard]
      summary: Expenses grouped by category
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySummary"
  /dashboard/cashflow:
    get:
      operationId: getDashboardCashflow
      tags: [dashboard]
      summary: Income vs expenses per month (last 6 months)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cashflow data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CashflowEntry"
  /dashboard/recent-transactions:
    get:
      operationId: getRecentTransactions
      tags: [dashboard]
      summary: Last 5 transactions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recent transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

  # Reports
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    Category:
      type: object
      required: [id, userId, name, type, icon, color, isActive]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [income, expense]
        icon:
          type: string
        color:
          type: string
        isActive:
          type: boolean
    CategoryInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [income, expense]
        icon:
          type: string
        color:
          type: string
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [income, expense]
        icon:
          type: string
        color:
          type: string
        isActive:
          type: boolean

    # Transaction schemas
    Transaction:
      type: object
      required: [id, userId, accountId, categoryId, type, amount, date, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        accountId:
          type: integer
        categoryId:
          type: integer
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        description:
          type: ["string", "null"]
        date:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        categoryName:
          type: ["string", "null"]
        categoryColor:
          type: ["string", "null"]
        categoryIcon:
          type: ["string", "null"]
        accountName:
          type: ["string", "null"]
...
    Budget:
      type: object
      required: [id, userId, categoryId, month, year, amount]
      properties:
        id:
          type: integer
        userId:
          type: integer
        categoryId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        amount:
          type: number
        categoryName:
          type: ["string", "null"]
        categoryColor:
          type: ["string", "null"]
        categoryIcon:
          type: ["string", "null"]
    BudgetInput:
      type: object
      required: [categoryId, month, year, amount]
      properties:
        categoryId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        amount:
          type: number
    BudgetUpdate:
      type: object
      properties:
        amount:
          type: number
    BudgetStatus:
      type: object
      required: [categoryId, categoryName, budgeted, spent, percentage]
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryColor:
          type: ["string", "null"]
        categoryIcon:
          type: ["string", "null"]
        budgeted:
          type: number
        spent:
          type: number
        percentage:
          type: number

    # Goal schemas
...
    DashboardSummary:
      type: object
      required: [totalBalance, monthlyIncome, monthlyExpense, monthlyBalance]
      properties:
        totalBalance:
          type: number
        monthlyIncome:
          type: number
        monthlyExpense:
          type: number
        monthlyBalance:
          type: number
        month:
          type: integer
        year:
          type: integer
    CategorySummary:
      type: object
      required: [categoryId, categoryName, total]
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryColor:
          type: ["string", "null"]
        categoryIcon:
          type: ["string", "null"]
        total:
          type: number
    CashflowEntry:
      type: object
      required: [month, year, income, expense]
      properties:
        month:
          type: integer
        year:
          type: integer
        label:
          type: string
        income:
          type: number
        expense:
          type: number

    # Report schemas
...