# lib/api-zod/src/generated/types/categorySummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sistema de Gestão Financeira Pessoal API
 * OpenAPI spec version: 0.1.0
 */

export interface CategorySummary {
  categoryId: number;
  categoryName: string;
  /** @nullable */
  categoryColor?: string | null;
  /** @nullable */
  categoryIcon?: string | null;
  total: number;
}